+++ packages/wallets/src/index.ts
@@ -5,7 +5,7 @@ export { createCrossmint, CrossmintWallets } from "./sdk";
 export { ApiClient as WalletsApiClient } from "./api";
 
 // Types
-export type { ShadowSignerStorage, ShadowSignerData } from "./signers/shadow-signer";
+export type { ShadowSignerStorage, ShadowSignerData } from "./signers/shadow-signer/utils";
 
 // Wallets
 export { Wallet } from "./wallets/wallet";
