+++ packages/wallets/src/signers/shadow-signer/solana-shadow-signer.ts
@@ -2,7 +2,7 @@ import { PublicKey } from "@solana/web3.js";
 import { ShadowSigner } from "./shadow-signer";
 import type { SolanaChain } from "@/chains/chains";
 import type { ExternalWalletInternalSignerConfig } from "../types";
-import type { ShadowSignerData } from "./index";
+import type { ShadowSignerData } from "./utils";
 import { SolanaExternalWalletSigner } from "../solana-external-wallet";
 
 export class SolanaShadowSigner extends ShadowSigner<SolanaChain> {
